Documents/Projects/FashionFit/FashionFitLocal/project/my_flask_app/requirements.txt
flask
flask-cors
pymongo
pytz
requests
python-dotenv
jwt
